pyproject.toml
[build-system]
requires = ["setuptools>=80.9", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airspace-viewer"
version = "0.1.0"
description = "A Python-based interactive airspace visualization tool for viewing OpenAir airspace files"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Simon Steiner", email = "simonsteiner@users.noreply.github.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Flask",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
keywords = ["airspace", "openair", "aviation", "gis", "visualization", "flask", "mapping"]
requires-python = ">=3.8"
dependencies = [
    "flask>=3.1.1",
    "openair-rs-py>=0.1.4",
    "simplekml>=1.3.6",
    "werkzeug>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.16.1",
    "pre_commit>=4.2.0",
    "pydocstyle>=6.3.0",
]

[project.urls]
Repository = "https://github.com/simonsteiner/airspace-viewer"

[project.scripts]
airspace-viewer = "app:main"
